ENH: on failure, the normalizer decorator either raise (default) or return a value
On failure,  @ normalizer without arguments (default) raises, while
@normalizer(return_or_failure=value) return a specified `value`.

This is useful, e.g. for

@-normalizer
def is_scalar(a: ArrayLike):
   ....

Consider is_scalar(int): the argument cannot be converted to tensor, so
the function returns False. But the actual failure is in the decorator,
so cannot do try-except in the body of the function!