Fix fill_diagonal
fill_diagonal had an implementation that assumed that `.ravel()` returns
a view onto the original tensor. This is not true, as there are many
tensors that cannot be flattened. `x.flat` in NumPy returns an iterator
that iterates the elements "in order". We do not have anything like that
in PyTorch. This PR implements this operation via some rather funky
advanced indexing. The `wrap` kwarg is... weird to say the least...